New database model

+++ app/src/main/java/com/example/shopping_list/entity/Basket.kt
@@ -1,7 +1,7 @@
 package com.example.shopping_list.entity
 
 interface Basket {
-    var idBasket: Int
+    var idBasket: Long
     var nameBasket: String
     var fillBasket: Boolean
 }
